refactor(cli): lift catalog/apis printers and split identity_claim (E6)
Move the catalog and apis pretty-printers into catalog_print.go and
apis_print.go (apis.go keeps its os.WriteFile spec-download path), and
carve the human claim command + claimErr out of identity.go into
identity_claim.go (matching the existing identity_claim_test.go).
Intra-package moves only.
